Added exit when cd fails to not delete local repo
I changed the `git clone` command to use a local repo, but it
failed for some reason and cloned into the current directory, then
it attempted to change to the directory it was supposed to have
cloned the git repo into, which failed leaving us in the original
directory containing the WP Plugin Template fork I was working on.
The next line in the shell script is `rm -rf .git` which deleted
all of the work I had just committed immediately before running
this command, thinking I'd be safe.
Adding `|| exit` to any Bash cd command will exit if the command
fails. It's a common idiom in Bash that shellcheck.net helped me
identify.